Confirm before deleting students/plans and before cancelling Registrations

+++ gympoint-web/src/pages/Plans/index.js
@@ -21,7 +21,11 @@ export default function Plans() {
   }
 
   async function handleDelete(id) {
-    dispatch(deletePlanRequest(id, loadPlans));
+    const result = window.confirm('Tem certeza disso?');
+
+    if (result === true) {
+      dispatch(deletePlanRequest(id, loadPlans));
+    }
   }
 
   useEffect(() => {
